Add CI testing for EnzymeCore
